Fix zuul pylint setup failures
The pylint zuul target will now be setup the same as the
py39 target.

The pip legacy resolver is still needed because not all
of the other STX dependencies being installed by tox
are aligned properly for python 3.9

Enforce the nodeset for pylint to be the same as py39

Tox 4 was enabled by:
https://review.opendev.org/c/zuul/zuul-jobs/+/866943

Tox 4 deprecated whitelist_externals.
Replace whitelist_externals with allowlist_externals

Partial-Bug: #1997255
Partial-Bug: #2000399
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I7d7ca4531600bd863a045d3daeb0be82fec34377